Introduce require_external(...) for shared externals.
