+++ core/CMakeLists.txt
@@ -14,109 +14,13 @@ if(BUILD_CORE)
   set(DEP_LIST "${DEP_LIST};BUILD_CORE BUILD_VISLIB" CACHE INTERNAL "")
   set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${CMAKE_CURRENT_SOURCE_DIR}/cmake)
 
-  # search for 3rd party libs
+  # Add threading.
   set(CMAKE_THREAD_PREFER_PTHREAD)
-
   find_package(Threads REQUIRED)
 
-  # include lua sources
-  add_subdirectory(lua)
-
-  # include ZMQ and its C++ bindings
-  set(ZMQ_VER "4_2_3")
-  string(REPLACE "_" "." ZMQ_TAG "v${ZMQ_VER}")
-  if(MSVC_IDE)
-    set(MSVC_TOOLSET "-${CMAKE_VS_PLATFORM_TOOLSET}")
-  else()
-    set(MSVC_TOOLSET "")
-  endif()
+  # Add expat.
   if(WIN32)
-    set(ZMQ_IMPORT_DEBUG "lib/libzmq${MSVC_TOOLSET}-mt-gd-${ZMQ_VER}.lib")
-    set(ZMQ_IMPORT_RELEASE "lib/libzmq${MSVC_TOOLSET}-mt-${ZMQ_VER}.lib")
-    set(ZMQ_DEBUG "bin/libzmq${MSVC_TOOLSET}-mt-gd-${ZMQ_VER}.dll")
-    set(ZMQ_RELEASE "bin/libzmq${MSVC_TOOLSET}-mt-${ZMQ_VER}.dll")
-  else()
-    include(GNUInstallDirs)
-    set(ZMQ_IMPORT_DEBUG "")
-    set(ZMQ_IMPORT_RELEASE "")
-    set(ZMQ_DEBUG "${CMAKE_INSTALL_LIBDIR}/libzmq.so")
-    set(ZMQ_RELEASE ${ZMQ_DEBUG})
-  endif()
-
-  add_external_project(libzmq_ext
-    GIT_REPOSITORY https://github.com/zeromq/libzmq.git
-    GIT_TAG ${ZMQ_TAG}
-    CMAKE_ARGS
-      -DZMQ_BUILD_TESTS=OFF)
-  add_external_library(libzmq SHARED
-    DEPENDS libzmq_ext
-    IMPORT_LIBRARY_DEBUG ${ZMQ_IMPORT_DEBUG}
-    IMPORT_LIBRARY_RELEASE ${ZMQ_IMPORT_RELEASE}
-    LIBRARY_DEBUG ${ZMQ_DEBUG}
-    LIBRARY_RELEASE ${ZMQ_RELEASE})
-
-  add_external_project(libcppzmq_ext
-    DEPENDS libzmq
-    GIT_REPOSITORY https://github.com/zeromq/cppzmq.git
-    GIT_TAG ${ZMQ_TAG}
-    CONFIGURE_COMMAND ""
-    BUILD_COMMAND ""
-    INSTALL_COMMAND ""
-    TEST_COMMAND "")
-  add_external_library(libcppzmq INTERFACE
-    DEPENDS libcppzmq_ext
-    INCLUDE_DIR "src/libcppzmq_ext/")
-
-  # Add zlib.
-  if(MSVC)
-    set(ZLIB_DEBUG "lib/zlibstaticd${CMAKE_STATIC_LIBRARY_SUFFIX}")
-    set(ZLIB_RELEASE "lib/zlibstatic${CMAKE_STATIC_LIBRARY_SUFFIX}")
-  else()
-    include(GNUInstallDirs)
-    #set(ZLIB_DEBUG "${CMAKE_INSTALL_LIBDIR}/${CMAKE_STATIC_LIBRARY_PREFIX}z${CMAKE_STATIC_LIBRARY_SUFFIX}")
-    #set(ZLIB_RELEASE "${CMAKE_INSTALL_LIBDIR}/${CMAKE_STATIC_LIBRARY_PREFIX}z${CMAKE_STATIC_LIBRARY_SUFFIX}")
-    set(ZLIB_DEBUG "lib/${CMAKE_STATIC_LIBRARY_PREFIX}z${CMAKE_STATIC_LIBRARY_SUFFIX}")
-    set(ZLIB_RELEASE "lib/${CMAKE_STATIC_LIBRARY_PREFIX}z${CMAKE_STATIC_LIBRARY_SUFFIX}")
-  endif()
-  add_external_project(zlib_ext
-    GIT_REPOSITORY https://github.com/madler/zlib.git
-    GIT_TAG "v1.2.11"
-    CMAKE_ARGS
-      -DCMAKE_POSITION_INDEPENDENT_CODE:BOOL=ON)
-  add_external_library(zlib STATIC
-    DEPENDS zlib_ext
-    INCLUDE_DIR "include"
-    LIBRARY_DEBUG ${ZLIB_DEBUG}
-    LIBRARY_RELEASE ${ZLIB_RELEASE})
-    
-  # Add libpng.
-  if(MSVC)
-    set(LIBPNG_DEBUG "lib/libpng16_staticd${CMAKE_STATIC_LIBRARY_SUFFIX}")
-    set(LIBPNG_RELEASE "lib/libpng16_static${CMAKE_STATIC_LIBRARY_SUFFIX}")
-  else()
-    include(GNUInstallDirs)
-    set(LIBPNG_DEBUG "${CMAKE_INSTALL_LIBDIR}/${CMAKE_STATIC_LIBRARY_PREFIX}png16${CMAKE_STATIC_LIBRARY_SUFFIX}")
-    set(LIBPNG_RELEASE "${CMAKE_INSTALL_LIBDIR}/${CMAKE_STATIC_LIBRARY_PREFIX}png16${CMAKE_STATIC_LIBRARY_SUFFIX}")
-  endif()
-  ExternalProject_Get_Property(zlib_ext INSTALL_DIR)
-  add_external_project(libpng_ext
-    GIT_REPOSITORY https://git.code.sf.net/p/libpng/code
-    GIT_TAG "v1.6.34"
-    DEPENDS zlib_ext
-    CMAKE_ARGS
-      -DPNG_SHARED=OFF
-      -DPNG_TESTS=OFF
-      -DCMAKE_POSITION_INDEPENDENT_CODE:BOOL=ON
-      -DCMAKE_PREFIX_PATH:PATH=${INSTALL_DIR})
-  add_external_library(libpng STATIC
-    DEPENDS libpng_ext
-    INCLUDE_DIR "include"
-    LIBRARY_DEBUG ${LIBPNG_DEBUG}
-    LIBRARY_RELEASE ${LIBPNG_RELEASE}
-    INTERFACE_LIBRARIES zlib)
-
-  if(WIN32)
-    option(USE_INTERNAL_EXPAT "Compiles the expat provided by megamol" ON)
+    option(USE_INTERNAL_EXPAT "Compiles the expat provided by MegaMol" ON)
     if(USE_INTERNAL_EXPAT)
       add_subdirectory(expat)
       set(EXPAT_LIBRARIES expat)
@@ -128,65 +32,18 @@ if(BUILD_CORE)
     find_package(EXPAT REQUIRED)
   endif()
 
-  # include zfp
-  if(WIN32)
-    set(ZFP_LIB "lib/zfp.lib")
-  else()
-    include(GNUInstallDirs)
-    set(ZFP_LIB "${CMAKE_INSTALL_LIBDIR}/libzfp.a")
-  endif()
+  # Add Lua.
+  add_subdirectory(lua)
 
-  add_external_project(zfp_ext
-    GIT_REPOSITORY https://github.com/LLNL/zfp.git
-    GIT_TAG "0.5.2"
-    CMAKE_ARGS
-      -DBUILD_SHARED_LIBS=OFF
-      -DBUILD_UTILITIES=OFF
-      -DBUILD_TESTING=OFF
-      -DZFP_WITH_ALIGNED_ALLOC=ON
-      -DZFP_WITH_CACHE_FAST_HASH=ON
-      -DCMAKE_BUILD_TYPE=Release)
-  add_external_library(zfp STATIC
-    DEPENDS zfp_ext
-    LIBRARY ${ZFP_LIB})
-
-  # include glm
-  add_external_project(glm_ext
-      GIT_REPOSITORY https://github.com/g-truc/glm.git
-      GIT_TAG "0.9.8"
-      CONFIGURE_COMMAND ""
-      BUILD_COMMAND ""
-      INSTALL_COMMAND ""
-      TEST_COMMAND ""
-      CMAKE_ARGS -DGLM_TEST_ENABLE=OFF)
-  add_external_library(glm INTERFACE
-    DEPENDS glm_ext
-    INCLUDE_DIR "src/glm_ext/")
-
-  # include glowl
-  add_external_project(glowl_ext
-    GIT_REPOSITORY https://github.com/invor/glowl.git
-    GIT_TAG "v0.1"
-    CONFIGURE_COMMAND ""
-    BUILD_COMMAND ""
-    INSTALL_COMMAND ""
-    TEST_COMMAND "")
-  add_external_library(glowl INTERFACE
-    DEPENDS glowl_ext
-    INCLUDE_DIR "src/glowl_ext/include")
-
-  #include nlohmann/json
-  add_external_project(json_ext
-      GIT_REPOSITORY https://github.com/azadkuh/nlohmann_json_release.git
-      GIT_TAG "v3.5.0"
-      CONFIGURE_COMMAND ""
-      BUILD_COMMAND ""
-      INSTALL_COMMAND ""
-      TEST_COMMAND ""
-      CMAKE_ARGS -DBUILD_TESTING=OFF)
-  add_external_library(json INTERFACE
-    DEPENDS json_ext
-    INCLUDE_DIR "src/json_ext/")
+  # Add externals.
+  require_external(libzmq)
+  require_external(libcppzmq)
+  require_external(zlib)
+  require_external(libpng)
+  require_external(zfp)
+  require_external(glm)
+  require_external(glowl)
+  require_external(json)
 
   file(GLOB_RECURSE public_header_files RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}" "include/*.h" "include/*.inl")
   file(GLOB_RECURSE source_files RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}" "src/*.cpp")
@@ -212,7 +69,7 @@ if(BUILD_CORE)
   # Target definition
   add_library(${PROJECT_NAME} SHARED ${public_header_files} ${header_files} ${shader_files} ${source_files} ${version_files})
 
-  # add libunwind
+  # Add libunwind.
   if(UNIX)
     option(USE_LIBUNWIND "use libunwind" OFF)
     if(USE_LIBUNWIND)
@@ -228,6 +85,7 @@ if(BUILD_CORE)
       endif()
     endif()
   endif()
+  
   target_compile_definitions(${PROJECT_NAME} PRIVATE MEGAMOLCORE_EXPORTS)
   target_include_directories(${PROJECT_NAME}
     PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include> "src" "include"
