--- plugins/mmstd_datatools/3rd/nanoflann/readme.txt
@@ -1,2 +1,0 @@
-https://github.com/jlblancoc/nanoflann
-BSD Licensed
