+++ plugins/mmstd_datatools/CMakeLists.txt
@@ -18,11 +18,6 @@ if(BUILD_MMSTD_DATATOOLS_PLUGIN)
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${OpenMP_CXX_FLAGS}")
   endif()
 
-  add_external_project(tinyply_ext
-  GIT_REPOSITORY https://github.com/ddiakopoulos/tinyply.git
-  GIT_TAG "2.1"
-  CMAKE_ARGS -DSHARED_LIB=true)
-
   if(WIN32)
     set(TNY_IMPORT_LIB "lib/tinyply.lib")
     set(TNY_IMPORT_LIB_DEBUG "lib/tinyplyd.lib")
@@ -36,24 +31,16 @@ if(BUILD_MMSTD_DATATOOLS_PLUGIN)
     set(TNY_LIB "lib/libtinyply.so")
   endif()
 
+  add_external_project(tinyply_ext
+    GIT_REPOSITORY https://github.com/ddiakopoulos/tinyply.git
+    GIT_TAG "2.1"
+    CMAKE_ARGS -DSHARED_LIB=true)
   add_external_library(tinyply SHARED
-  DEPENDS tinyply_ext
-  IMPORT_LIBRARY_DEBUG ${TNY_IMPORT_LIB_DEBUG}
-  IMPORT_LIBRARY_RELEASE ${TNY_IMPORT_LIB}
-  LIBRARY_DEBUG ${TNY_LIB_DEBUG}
-  LIBRARY_RELEASE ${TNY_LIB})
-
-  # Add Eigen.
-  add_external_project(Eigen_ext
-    GIT_REPOSITORY https://github.com/eigenteam/eigen-git-mirror.git
-    GIT_TAG "3.3.4"
-    CONFIGURE_COMMAND ""
-    BUILD_COMMAND ""
-    INSTALL_COMMAND ""
-    TEST_COMMAND "")
-  add_external_library(Eigen INTERFACE
-    DEPENDS Eigen_ext
-    INCLUDE_DIR "src/Eigen_ext")
+    DEPENDS tinyply_ext
+    IMPORT_LIBRARY_DEBUG ${TNY_IMPORT_LIB_DEBUG}
+    IMPORT_LIBRARY_RELEASE ${TNY_IMPORT_LIB}
+    LIBRARY_DEBUG ${TNY_LIB_DEBUG}
+    LIBRARY_RELEASE ${TNY_LIB})
 
   add_external_project(mmpld_io_ext
     GIT_REPOSITORY https://github.com/UniStuttgart-VISUS/mmpld_io.git
@@ -64,6 +51,10 @@ if(BUILD_MMSTD_DATATOOLS_PLUGIN)
   add_external_library(mmpld_io INTERFACE
     DEPENDS mmpld_io_ext
     INCLUDE_DIR "src/mmpld_io_ext/include")
+
+  # Add externals.
+  require_external(Eigen)
+  require_external(nanoflann)
 
   file(GLOB_RECURSE public_header_files RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}" "include/*.h")
   file(GLOB_RECURSE source_files RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}" "src/*.cpp" "3rd/min_sphere_of_spheres/*.cpp")
@@ -95,8 +86,8 @@ if(BUILD_MMSTD_DATATOOLS_PLUGIN)
   set_target_properties(${PROJECT_NAME} PROPERTIES SUFFIX ".mmplg")
   target_compile_definitions(${PROJECT_NAME} PRIVATE ${EXPORT_NAME}_EXPORTS)
   target_include_directories(${PROJECT_NAME} PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include> "include" "src"
-    "3rd/nanoflann" "3rd/min_sphere_of_spheres")
-  target_link_libraries(${PROJECT_NAME} PRIVATE core geometry_calls tinyply Eigen mmpld_io)
+    "3rd/min_sphere_of_spheres")
+  target_link_libraries(${PROJECT_NAME} PRIVATE core geometry_calls tinyply Eigen nanoflann mmpld_io)
 
   if(MPI_C_FOUND)
     target_include_directories(${PROJECT_NAME} PUBLIC ${MPI_C_INCLUDE_PATH})
