+++ plugins/mmstd_datatools/src/ParticleNeighborhood.h
@@ -16,7 +16,7 @@
 #include "mmcore/moldyn/MultiParticleDataCall.h"
 #include "PointcloudHelpers.h"
 #include <vector>
-#include "nanoflann.hpp"
+#include <nanoflann.hpp>
 
 namespace megamol {
 namespace stdplugin {
