+++ plugins/mmstd_datatools/src/ParticleVelocities.cpp
@@ -8,7 +8,7 @@
 #include "ParticleVelocities.h"
 #include "mmcore/param/BoolParam.h"
 #include "mmcore/param/FloatParam.h"
-#include "nanoflann.hpp"
+#include <nanoflann.hpp>
 #include "vislib/sys/Log.h"
 #include "vislib/math/ShallowVector.h"
 #include "vislib/math/ShallowPoint.h"
