+++ plugins/mmstd_datatools/src/table/TableToLines.cpp
@@ -14,11 +14,45 @@
 #define M_PI 3.14159265358979323846
 #endif
 
-#include "nanoflann.hpp"
-#include "utils.h"
+#include <nanoflann.hpp>
 
 using namespace megamol::stdplugin::datatools;
 using namespace megamol;
+
+
+/*
+ * THIS IS THE APEX OF SHIT and a non-quality copy from nanoflann/examples/utils.h
+ * TODO: Replace it with a proper adapter instead of creating a copy to index data!
+ */
+template <typename T>
+struct PointCloud
+{
+	struct Point
+	{
+		T  x,y,z;
+	};
+
+	std::vector<Point>  pts;
+
+	// Must return the number of data points
+	inline size_t kdtree_get_point_count() const { return pts.size(); }
+
+	// Returns the dim'th component of the idx'th point in the class:
+	// Since this is inlined and the "dim" argument is typically an immediate value, the
+	//  "if/else's" are actually solved at compile time.
+	inline T kdtree_get_pt(const size_t idx, const size_t dim) const
+	{
+		if (dim == 0) return pts[idx].x;
+		else if (dim == 1) return pts[idx].y;
+		else return pts[idx].z;
+	}
+
+	// Optional bounding-box computation: return false to default to a standard bbox computation loop.
+	//   Return true if the BBOX was already computed by the class and returned in "bb" so it can be avoided to redo it again.
+	//   Look at bb.size() to find out the expected dimensionality (e.g. 2 or 3 for point clouds)
+	template <class BBOX>
+	bool kdtree_get_bbox(BBOX& /* bb */) const { return false; }
+};
 
 /*
  * FloattableToLines::FloattableToLines
