+++ vislib/tests/clusterTest/CMakeLists.txt
@@ -3,8 +3,8 @@
 #
 cmake_minimum_required(VERSION 3.0)
 
-if (NOT DEFINED BUILDING_VISLIB_PROJECT)
-	message(FATAL_ERROR "This CMakefile cannot be processed independently.")
+if(NOT DEFINED BUILDING_VISLIB_PROJECT)
+  message(FATAL_ERROR "This CMakefile cannot be processed independently.")
 endif()
 
 #input file
